+++ src/assets/activity/ActivityAsset.php
@@ -1,0 +1,45 @@
+<?php
+/**
+ * @link https://craftcms.com/
+ * @copyright Copyright (c) Pixel & Tonic, Inc.
+ * @license MIT
+ */
+
+namespace craft\webhooks\assets\activity;
+
+use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
+use craft\web\View;
+
+/**
+ * Webhooks index asset bundle
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 1.0
+ */
+class ActivityAsset extends AssetBundle
+{
+    public $sourcePath = __DIR__ . '/dist';
+
+    public $depends = [
+        CpAsset::class,
+    ];
+
+    public $js = [
+        'js/Activity.js',
+    ];
+
+    /**
+     * @inheritdoc
+     */
+    public function registerAssetFiles($view)
+    {
+        parent::registerAssetFiles($view);
+
+        if ($view instanceof View) {
+            $view->registerTranslations('webhooks', [
+                'Are you sure you want to resend this request?',
+            ]);
+        }
+    }
+}