+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 ### Added
+- Webhooks now logs requests, and it’s possible to view them from a new “Activity” page within the plugin.
+- Added new `maxAttempts` and `attemptDelay` settings, which can be set from `config/webhooks.php`.
 - The Sender Class and Event Name webhook settings now show suggestions based on the available classes and events.
 - Webhooks can now have custom payloads. ([#3](https://github.com/craftcms/webhooks/pull/3))